Stop caching `season` method
Stop caching `season` method because the initial request that
fetches all the seasons can be slow depending on the includes. For example,
calling `season` with include `results.odds` for the first time
would fetch all seasons with all the results of a season and the
odds for each result (a result is a completed fixture). Event with free
plan that gives all seasons of only two leagues this was an extremely
slow request.