Stop using "next" link from pagination object
If requested data spans multiple pages, the SportMonks API response
includes a pagination object. This object contains link
to the next page. However, this link does not include the query
string parameters from the original request. As such this link should
not be used. In this commit, we refactor `_http_get` not to use this
link and instead use the `current_page` and `total_pages` counters to
fetch other pages.