Stop caching
Delete property `_callables_cached_objects` and method `_lookup_table`
which were used to cache the return values of 4 methods: `league`,
`country`, `bookmaker`, and `continent`.

If caching is required then the caller must implement the cache. For
example, the `functools.lru_cache` decorator could be used to cache the
return values.