Change type hint of param `includes` to Iterable
Since no caching is done anymore, we don't need to have immutable data
types in the function signature as was required by `functools.lru_cache`
decorator. This means `includes` does not have to be a tuple (immutable)
anymore. Any iterable (mutable and immutable) should work now. Adjust
the type hint for this.