Fix bug with incorrect URL construction
We were constructing URLs with os.path.join which uses forward slash as
separator on Linux. However, on Windows this separator is a backslash
and so the constructed URL is incorrect on Windows. Swtich to
constructing URLs with urllib.parse.urljoin, which has its own "quirks".
For example the URL must end with trailing slash or else during join the
last part will be silently left out.