Refactor `_http_get` and its `endpoint` parameter
Refactor BaseApiV2._http_get: it's endpoint parameter now accepts a
single element or a list of elements that comprise an endpoint. From
these, _full_url is called to build up proper url where the request
should go to. In the end, this decreases amount of code because now we
do not need to supply a url-string to `endpoint`.