Refactor to please mypy
Mypy assign a type to a variable based on the first assignment
statement. It allows a variable to be downcasted later on: e.g. from
list of objects to a list of strings, but not upcasted: from string to a
list of objects. So we need to perform some gymnastics here.

Additionally refactor the type aliases EndpointPart and EndpointParts to
please mypy: make EndpointPart an object not a union of basic types.
When mypy sees a list of [str, str, int] it seems to consider it to be
List[object] instead of List[Union[str, int]].